# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NyayaSetu Legal Operating System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface EmptyInput { [key: string]: unknown }

export interface User {
  id: number;
  fullName: string;
  email: string;
  /** @nullable */
  mobile?: string | null;
  /** @nullable */
  aadhaar?: string | null;
  /** @nullable */
  address?: string | null;
  preferredLanguage: string;
  /** @nullable */
  profilePhoto?: string | null;
  role: string;
  createdAt: string;
}

export interface AuthResult {
  user: User;
  token: string;
}

export interface RegisterInput {
  fullName: string;
  email: string;
  mobile?: string;
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
  rememberMe?: boolean;
}

export interface ProfileUpdate {
  fullName?: string;
  mobile?: string;
  aadhaar?: string;
  address?: string;
  preferredLanguage?: string;
  profilePhoto?: string;
}

export interface PasswordChangeInput {
  currentPassword: string;
  newPassword: string;
}

export interface Complaint {
  id: number;
  type: string;
  title: string;
  content: string;
  status: string;
  answers?: string;
  createdAt: string;
  updatedAt: string;
}

export interface Document {
  id: number;
  name: string;
  originalName: string;
  /** @nullable */
  category?: string | null;
  mimeType: string;
  size: number;
  filePath: string;
  createdAt: string;
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  createdAt: string;
}

export interface Scheme {
  id: number;
  title: string;
  description: string;
  category: string;
  eligibility: string;
  benefits: string;
  requiredDocuments: string[];
  applicationProcess: string;
  /** @nullable */
  officialLink?: string | null;
  ministry: string;
  createdAt: string;
}

export interface RightsArticle {
  id: number;
  title: string;
  summary: string;
  content: string;
  category: string;
  relevantLaws: string[];
  legalProvisions: string[];
  governmentResources: string[];
  createdAt: string;
}

export interface DashboardStats {
  welcomeMessage: string;
  totalDocuments: number;
  totalComplaints: number;
  activeJourneys: number;
  unreadNotifications: number;
  recentComplaints: Complaint[];
  recentDocuments: Document[];
  recentActivity: ActivityItem[];
  featuredSchemes: Scheme[];
  legalRightsHighlight: RightsArticle[];
}

export interface RightsCategory {
  id: string;
  name: string;
  icon: string;
  count: number;
}

export interface RightsListResponse {
  articles: RightsArticle[];
  total: number;
  page: number;
  totalPages: number;
}

export interface ComplaintQuestion {
  key: string;
  label: string;
  type: string;
  required: boolean;
  /** @nullable */
  options?: string[] | null;
}

export interface ComplaintTemplate {
  id: number;
  type: string;
  title: string;
  description: string;
  questions: ComplaintQuestion[];
}

export interface ComplaintInput {
  type: string;
  title: string;
  answers?: string;
  content: string;
}

export interface ComplaintUpdate {
  title?: string;
  content?: string;
  status?: string;
}

export interface DocumentUpdate {
  name?: string;
  category?: string;
}

export interface JourneyStep {
  stepNumber: number;
  title: string;
  description: string;
  /** @nullable */
  tips?: string | null;
}

export interface Journey {
  id: number;
  title: string;
  description: string;
  category: string;
  estimatedDays: number;
  eligibility: string;
  requiredDocuments: string[];
  applicableLaws: string[];
  steps: JourneyStep[];
}

export interface JourneyProgress {
  journeyId: number;
  currentStep: number;
  completedSteps: number[];
  completionPercent: number;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
}

export interface JourneyProgressUpdate {
  currentStep: number;
  completedSteps: number[];
}

export interface SchemesListResponse {
  schemes: Scheme[];
  total: number;
  page: number;
  totalPages: number;
}

export interface LearningContent {
  id: number;
  title: string;
  summary: string;
  content: string;
  type: string;
  tags: string[];
  isBookmarked: boolean;
  createdAt: string;
}

export interface BookmarkInput {
  bookmarked: boolean;
}

export interface LegalAidProvider {
  id: number;
  name: string;
  type: string;
  city: string;
  state: string;
  /** @nullable */
  pincode?: string | null;
  address: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  website?: string | null;
  practiceAreas: string[];
  languages: string[];
  /** @nullable */
  rating?: number | null;
  description: string;
}

export interface ChatSession {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
  messageCount: number;
}

export interface ChatSessionInput {
  title: string;
}

export interface ChatMessage {
  id: number;
  sessionId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface ChatMessageInput {
  message: string;
}

export interface Notification {
  id: number;
  type: string;
  title: string;
  message: string;
  isRead: boolean;
  createdAt: string;
}

export interface NotificationReadInput {
  isRead: boolean;
}

export interface SearchResults {
  rights: RightsArticle[];
  schemes: Scheme[];
  complaints: Complaint[];
  documents: Document[];
  learning: LearningContent[];
}

export interface AdminLoginInput {
  email: string;
  password: string;
}

export interface CountItem {
  label: string;
  count: number;
}

export interface AdminStats {
  totalUsers: number;
  totalDocuments: number;
  totalComplaints: number;
  totalSessions: number;
  newUsersThisMonth: number;
  activeUsersToday: number;
  topComplaintTypes: CountItem[];
  topRightsCategories: CountItem[];
}

export interface AdminUsersResponse {
  users: User[];
  total: number;
  page: number;
  totalPages: number;
}

export type ListRightsParams = {
category?: string;
q?: string;
page?: number;
limit?: number;
};

export type ListDocumentsParams = {
q?: string;
category?: string;
};

export type ListSchemesParams = {
q?: string;
category?: string;
page?: number;
};

export type ListLearningContentParams = {
type?: string;
q?: string;
};

export type ListLegalAidParams = {
city?: string;
state?: string;
type?: string;
area?: string;
q?: string;
};

export type GlobalSearchParams = {
q: string;
};

export type AdminListUsersParams = {
q?: string;
page?: number;
};