# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NyayaSetu Legal Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated

  # USER PROFILE
...
  /complaints:
    get:
      operationId: listComplaints
      tags: [complaints]
      summary: List user's complaints
      responses:
        "200":
          description: Complaints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Complaint"
    post:
      operationId: createComplaint
      tags: [complaints]
      summary: Create a new complaint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplaintInput"
      responses:
        "201":
          description: Created complaint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
  /complaints/{id}:
    get:
      operationId: getComplaint
      tags: [complaints]
      summary: Get a complaint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Complaint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
    patch:
      operationId: updateComplaint
      tags: [complaints]
      summary: Update complaint content
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplaintUpdate"
      responses:
        "200":
          description: Updated complaint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
    delete:
      operationId: deleteComplaint
      tags: [complaints]
      summary: Delete a complaint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # DOCUMENT VAULT
...
  /chat/sessions:
    get:
      operationId: listChatSessions
      tags: [chat]
      summary: List user's chat sessions
      responses:
        "200":
          description: Sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatSession"
    post:
      operationId: createChatSession
      tags: [chat]
      summary: Create a new chat session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatSessionInput"
      responses:
        "201":
          description: Created session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSession"
  /chat/sessions/{id}/messages:
    get:
      operationId: getChatMessages
      tags: [chat]
      summary: Get messages for a session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message and get AI response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"

  # NOTIFICATIONS
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersResponse"
  /admin/users/{id}:
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      summary: Delete a user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    AuthResult:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    User:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        mobile:
          type: ["string", "null"]
        aadhaar:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        preferredLanguage:
          type: string
        profilePhoto:
          type: ["string", "null"]
        role:
          type: string
        createdAt:
          type: string
      required: [id, fullName, email, preferredLanguage, role, createdAt]
    RegisterInput:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        mobile:
          type: string
        password:
          type: string
      required: [fullName, email, password]
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        rememberMe:
          type: boolean
      required: [email, password]
...
    Complaint:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        title:
          type: string
        content:
          type: string
        status:
          type: string
        answers:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, type, title, content, status, createdAt, updatedAt]
    ComplaintInput:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        answers:
          type: string
        content:
          type: string
      required: [type, title, content]
    ComplaintUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        status:
          type: string
...
    ChatSession:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        messageCount:
          type: integer
      required: [id, title, createdAt, updatedAt, messageCount]
    ChatSessionInput:
      type: object
      properties:
        title:
          type: string
      required: [title]
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
      required: [id, sessionId, role, content, createdAt]
    ChatMessageInput:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    AdminLoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
...
    AdminUsersResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [users, total, page, totalPages]